BizSenseAI/requirements.txt
flask
google-genai
pandas
werkzeug
python-dotenv
